Fixes for optimizing with stock piece quantities
Sometimes the optimizer would fail in cases it should give a solution
when the stock pieces were limited with a quantity.

To fix this, each unit keeps a list of cut pieces it couldn't use. This
allows us to continue to find a better solution, instead of failing when
a fit is not found for a cut piece.